+++ cli/src/login.rs
@@ -1,13 +1,17 @@
-use anyhow::{Context, Error};
+use anyhow::{bail, Context, Error, Result};
 use inquire::{Select, Text};
 use serde::{Deserialize, Serialize};
-use std::thread;
+use std::io::IsTerminal;
 use std::time::Duration;
+use std::{io, thread};
 use tracing::info;
 
 use crate::{
     invocation_context::{context, init_context},
-    utils::auth::{host_validator, token_validator, CredentialProvider, HomeDirProvider, Token},
+    utils::auth::{
+        env_id_validator, host_validator, token_validator, CredentialProvider, HomeDirProvider,
+        Token,
+    },
 };
 
 #[derive(Debug, Deserialize)]
@@ -31,14 +35,14 @@ struct PollResponse {
     project_id: Option<String>,
 }
 
-pub fn login(host_override: Option<String>) -> Result<(), Error> {
+pub fn login(host_override: Option<String>) -> Result<()> {
+    if !io::stdout().is_terminal() {
+        bail!("Failed to login. If you are running on a CI, skip this step and use POSTHOG_CLI_HOST, POSTHOG_CLI_ENV_ID, POSTHOG_CLI_TOKEN env variables when running commands")
+    }
     login_with_use_cases(host_override, vec!["schema", "error_tracking"])
 }
 
-pub fn login_with_use_cases(
-    host_override: Option<String>,
-    use_cases: Vec<&str>,
-) -> Result<(), Error> {
+pub fn login_with_use_cases(host_override: Option<String>, use_cases: Vec<&str>) -> Result<()> {
     let host = if let Some(override_host) = host_override {
         // Strip trailing slashes to avoid double slashes in URLs
         override_host.trim_end_matches('/').to_string()
@@ -267,8 +271,9 @@ fn manual_login() -> Result<(), Error> {
         .with_validator(host_validator)
         .prompt()?;
 
-    let env_id =
-        Text::new("Enter your project ID (the number in your PostHog homepage URL)").prompt()?;
+    let env_id = Text::new("Enter your project ID (the number in your PostHog homepage URL)")
+        .with_validator(env_id_validator)
+        .prompt()?;
 
     let token = Text::new(
         "Enter your personal API token",