+++ posthog/hogql_queries/document_embeddings_query_runner.py
@@ -186,7 +186,8 @@ class DocumentEmbeddingsQueryRunner(AnalyticsQueryRunner[DocumentSimilarityQuery
                                     ),
                                 ]
                             ),
-                            ast.And(
+                            # If the document type or product is different - compare all renderings
+                            ast.Or(
                                 exprs=[
                                     ast.CompareOperation(
                                         op=ast.CompareOperationOp.NotEq,