+++ posthog/schema_helpers.py
@@ -109,6 +109,8 @@ def to_dict(query: BaseModel) -> dict:
                         "movingAverageIntervals",
                         "stacked",
                         "detailedResultsAggregationType",
+                        "showFullUrls",
+                        "selectedInterval",
                     ]
                 }